utils/mqtt_adapter/requirements.txt
loguru~=0.5.3
pyserial~=3.4
PyYAML~=5.3.1
ujson~=4.0.1